fix(core): Correctly read constants now
The issue was i was reading a unsigned 32 bit number instead of a 64 bit
double

Okay yeah i also did some changing towards the string reading.

String reading now pops the last item on the vec (the null byte) so i
wont have to deal with the null byte later when making an actual
frontend, lmao.